refactor: Separate file watcher
